PLAT-995 | Extract lesson add form
